Move client code in common into its own source set
