+++ common/src/client/java/io/github/gaming32/bingo/client/config/BoardCorner.java
