+++ common/src/client/java/io/github/gaming32/bingo/client/platform/event/ClientEvents.java
@@ -1,7 +1,7 @@
-package io.github.gaming32.bingo.platform.event;
+package io.github.gaming32.bingo.client.platform.event;
 
+import io.github.gaming32.bingo.platform.event.Event;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.player.LocalPlayer;
 
 import java.util.function.Consumer;
@@ -12,14 +12,4 @@ public class ClientEvents {
     public static final Event<Consumer<LocalPlayer>> PLAYER_QUIT = new Event<>();
     public static final Event<Consumer<Minecraft>> CLIENT_TICK_START = new Event<>();
     public static final Event<Consumer<Minecraft>> CLIENT_TICK_END = new Event<>();
-
-    @FunctionalInterface
-    public interface KeyReleaseHandler {
-        boolean onKeyReleased(Screen screen, int keyCode, int scanCode, int modifiers);
-    }
-
-    @FunctionalInterface
-    public interface MouseReleaseHandler {
-        boolean onMouseReleased(Screen screen, double mouseX, double mouseY, int modifiers);
-    }
 }
