+++ common/src/client/java/io/github/gaming32/bingo/client/platform/event/MouseReleaseHandler.java
@@ -1,0 +1,8 @@
+package io.github.gaming32.bingo.client.platform.event;
+
+import net.minecraft.client.gui.screens.Screen;
+
+@FunctionalInterface
+public interface MouseReleaseHandler {
+    boolean onMouseReleased(Screen screen, double mouseX, double mouseY, int modifiers);
+}
