+++ common/src/client/java/io/github/gaming32/bingo/mixin/common/client/CameraAccessor.java
