+++ common/src/client/resources/bingo-common-client.mixins.json
@@ -1,0 +1,16 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "io.github.gaming32.bingo.mixin.common.client",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "CameraAccessor",
+    "MixinGameRenderer",
+    "MixinGui",
+    "MixinKeyMapping",
+    "MixinRootSpectatorMenuCategory"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
