+++ fabric/src/main/resources/bingo-fabric.mixins.json
